feat: refresh dashboard v3 layout
