+++ apps/web/src/components/dashboard-v3/MissionsSection.tsx
@@ -2,6 +2,7 @@ import { useMemo } from 'react';
 import { useRequest } from '../../hooks/useRequest';
 import { getTasks, type UserTask } from '../../lib/api';
 import { asArray } from '../../lib/safe';
+import { Card } from '../ui/Card';
 
 interface MissionsSectionProps {
   userId: string;
@@ -15,14 +16,15 @@ export function MissionsSection({ userId }: MissionsSectionProps) {
   }, [data]);
 
   return (
-    <section className="rounded-3xl border border-white/10 bg-white/5 p-6 text-sm text-text backdrop-blur">
-      <header className="mb-4 flex flex-wrap items-center justify-between gap-3 text-white">
-        <h3 className="text-lg font-semibold">🗂️ Misiones</h3>
-        <span className="rounded-full border border-white/10 bg-white/10 px-3 py-1 text-xs uppercase tracking-wide text-text-muted">
-          Lectura solamente
+    <Card
+      title="🗂️ Misiones"
+      subtitle="Lectura solamente"
+      rightSlot={
+        <span className="inline-flex items-center gap-1 rounded-full border border-white/10 bg-white/5 px-2.5 py-1 text-xs text-slate-200">
+          Próximamente acciones
         </span>
-      </header>
-
+      }
+    >
       {status === 'loading' && <div className="h-32 w-full animate-pulse rounded-2xl bg-white/10" />}
 
       {status === 'error' && <p className="text-sm text-rose-300">No pudimos listar las misiones.</p>}
@@ -31,13 +33,13 @@ export function MissionsSection({ userId }: MissionsSectionProps) {
         <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
           {tasks.slice(0, 6).map((task) => (
             <article key={task.task_id} className="space-y-2 rounded-2xl border border-white/10 bg-white/5 p-4">
-              <h4 className="font-semibold text-white">🎯 {task.task}</h4>
-              <p className="text-xs text-text-muted">Pilar: {task.pillar_id ?? '—'}</p>
-              <p className="text-xs text-text-muted">XP base: {task.xp_base}</p>
-              <p className="text-xs text-text-muted">Constancia requerida: próximamente</p>
+              <h4 className="font-semibold text-slate-100">🎯 {task.task}</h4>
+              <p className="text-xs text-slate-400">Pilar: {task.pillar_id ?? '—'}</p>
+              <p className="text-xs text-slate-400">XP base: {task.xp_base}</p>
+              <p className="text-xs text-slate-400">Constancia requerida: próximamente</p>
               <button
                 type="button"
-                className="w-full rounded-full border border-white/10 bg-white/5 px-3 py-2 text-xs font-semibold uppercase tracking-wide text-text"
+                className="w-full rounded-full border border-white/10 bg-white/5 px-3 py-2 text-xs font-semibold uppercase tracking-[0.18em] text-slate-200"
                 disabled
               >
                 Activar
@@ -48,12 +50,12 @@ export function MissionsSection({ userId }: MissionsSectionProps) {
       )}
 
       {status === 'success' && tasks.length === 0 && (
-        <p className="text-sm text-text-muted">Tu base todavía no tiene misiones configuradas.</p>
+        <p className="text-sm text-slate-400">Tu base todavía no tiene misiones configuradas.</p>
       )}
 
-      <p className="mt-4 text-xs text-text-muted">
+      <p className="mt-4 text-xs text-slate-400">
         Mientras el endpoint de misiones dedicado no esté disponible reutilizamos las tareas activas como referencia visual.
       </p>
-    </section>
+    </Card>
   );
 }
