+++ apps/web/src/components/dashboard-v3/ProfileCard.tsx
@@ -1,3 +1,5 @@
+import { CardSection } from '../ui/Card';
+
 interface ProfileCardProps {
   imageUrl?: string | null;
 }
@@ -6,7 +8,7 @@ export function ProfileCard({ imageUrl }: ProfileCardProps) {
   const hasImage = typeof imageUrl === 'string' && imageUrl.trim().length > 0;
 
   return (
-    <section className="rounded-3xl border border-white/10 bg-white/5 p-6 backdrop-blur">
+    <CardSection aria-label="Perfil">
       <div className="aspect-[5/6] w-full">
         {hasImage ? (
           <img
@@ -15,11 +17,11 @@ export function ProfileCard({ imageUrl }: ProfileCardProps) {
             className="h-full w-full rounded-2xl object-cover shadow-lg"
           />
         ) : (
-          <div className="flex h-full w-full items-center justify-center rounded-2xl bg-white/10 text-5xl text-white/80 shadow-lg">
+          <div className="flex h-full w-full items-center justify-center rounded-2xl bg-white/10 text-5xl text-slate-200">
             👤
           </div>
         )}
       </div>
-    </section>
+    </CardSection>
   );
 }
