+++ apps/web/src/components/layout/Navbar.tsx
@@ -25,18 +25,6 @@ export function Navbar() {
         <div className="flex flex-1 items-center justify-end gap-3">
           <nav className="hidden items-center gap-2 md:flex">
             <NavLink
-              to="/dashboard"
-              className={({ isActive }) =>
-                `rounded-lg border px-3 py-2 text-xs font-semibold uppercase tracking-wide transition ${
-                  isActive
-                    ? 'border-white/40 bg-white/20 text-white'
-                    : 'border-white/10 bg-white/5 text-text hover:border-white/20 hover:bg-white/10'
-                }`
-              }
-            >
-              Dashboard clásico
-            </NavLink>
-            <NavLink
               to="/dashboard-v3"
               className={({ isActive }) =>
                 `rounded-lg border px-3 py-2 text-xs font-semibold uppercase tracking-wide transition ${
